Fixed precision issues with numeric comparisons

+++ src/main/java/io/github/redstoneparadox/paradoxconfig/util/Operators.kt
@@ -6,11 +6,11 @@ package io.github.redstoneparadox.paradoxconfig.util
 
 internal operator fun Number.compareTo(other: Number): Int {
     return when (this) {
-        is Byte -> this.compareTo(other.toDouble())
-        is Short -> this.compareTo(other.toDouble())
-        is Int -> this.compareTo(other.toDouble())
-        is Long -> this.compareTo(other.toDouble())
-        is Float -> this.compareTo(other.toDouble())
+        is Byte -> this.compareTo(other.toLong())
+        is Short -> this.compareTo(other.toLong())
+        is Int -> this.compareTo(other.toLong())
+        is Long -> this.compareTo(other.toLong())
+        is Float -> this.compareTo(other.toFloat())
         is Double -> this.compareTo(other.toDouble())
         else -> throw Exception("Something strange happened!")
     }
